WORKING: commit before removing any additional unnecessary variables

+++ final_project/lib/Battery/src/Battery.h
@@ -5,7 +5,7 @@
 #include "States.h"
 
 float Battery_getBatteryVoltage(void);
-uint8_t Battery_getBatteryPercentage(uint8_t batteryVoltage);
+uint8_t Battery_getBatteryPercentage(float batteryVoltage);
 
 bool Battery_getIsStablePowerSupply();
 uint8_t Battery_getChargeStatus();
